changed arg to double instead of float

+++ Program.cs
@@ -3,6 +3,6 @@ using static RandLib.CategoricalDist.CategoricalDist;
 
 //tests
 
-string result = flip((float)0.5) ? "heads" : "tails"; 
+string result = flip(0.5) ? "heads" : "tails"; 
 
 Console.WriteLine("flipped and got " + result + ".");
